[mission_manager] restructure node
