+++ ros_ws/src/ground_control_station/mission_manager/README.md
@@ -5,5 +5,5 @@ It assigned agents to search or track. For search it divides the search space.
 
 Debugging this node
 ```
-ros2 run mission_manager mission_manager_node
+ros2 run --prefix 'gdb -ex run --args'  mission_manager mission_manager_node
 ```
